feat(tasks): let new spaces link repositories during creation
Extracts the repository chips + add-picker into a shared, controlled RepositoriesField (used by the context page) and adds it as an optional "Repositories" step in the create-space dialog, styled like the existing name/description fields. Repos chosen there are linked to the space right after it's created; the field is skippable and the space still creates without any.

Generated-By: PostHog Code
Task-Id: 7d8b55e0-f9ff-45c3-acbc-155cae3cc474
