+++ products/desktop/packages/ui/src/features/canvas/components/CreateChannelModal.tsx
@@ -15,9 +15,11 @@ import {
   Textarea,
 } from "@posthog/quill";
 import { ANALYTICS_EVENTS } from "@posthog/shared/analytics-events";
+import { RepositoriesField } from "@posthog/ui/features/canvas/components/RepositoriesField";
 import { useChannelMutations } from "@posthog/ui/features/canvas/hooks/useChannels";
 import { useChannelsLayout } from "@posthog/ui/features/canvas/hooks/useChannelsLayout";
 import { useGenerateContext } from "@posthog/ui/features/canvas/hooks/useGenerateContext";
+import { useUpdateTaskChannelRepositories } from "@posthog/ui/features/canvas/hooks/useTaskChannels";
 import { toast } from "@posthog/ui/primitives/toast";
 import { track } from "@posthog/ui/shell/analytics";
 import { useNavigate } from "@tanstack/react-router";
@@ -56,9 +58,14 @@ export function CreateChannelModal({
   const spacesLayout = useChannelsLayout();
   const { createChannel, isCreating } = useChannelMutations();
   const { generate, isStarting } = useGenerateContext();
+  const linkRepositories = useUpdateTaskChannelRepositories();
   const navigate = useNavigate();
   const [name, setName] = useState("");
   const [description, setDescription] = useState("");
+  // Repositories to link to the new space, chosen in the optional step. They
+  // must share one GitHub integration, tracked alongside the selection.
+  const [repositories, setRepositories] = useState<string[]>([]);
+  const [repoIntegration, setRepoIntegration] = useState<number | null>(null);
   // Create mode's step. Describe mode has no name step, so it starts past it.
   const [step, setStep] = useState<"name" | "describe">("name");
 
@@ -71,6 +78,8 @@ export function CreateChannelModal({
     if (open) {
       setName("");
       setDescription("");
+      setRepositories([]);
+      setRepoIntegration(null);
       setStep("name");
     }
   }
@@ -80,7 +89,7 @@ export function CreateChannelModal({
   const remaining = MAX_CONTEXT_NAME_LENGTH - name.length;
   const nameError = isDescribeMode ? null : validateChannelName(trimmedName);
 
-  const busy = isCreating || isStarting;
+  const busy = isCreating || isStarting || linkRepositories.isPending;
   const canAdvance = !busy && !!trimmedName && !nameError;
   // "Create" seeds the plan session, so it needs the description; "Skip" is the
   // way through without one.
@@ -125,6 +134,23 @@ export function CreateChannelModal({
         description: error instanceof Error ? error.message : String(error),
       });
       return;
+    }
+
+    // Link the chosen repositories to the fresh space. The space already
+    // exists, so a failure here only warns — the user can retry from its
+    // context page.
+    if (repositories.length > 0) {
+      try {
+        await linkRepositories.mutateAsync({
+          channelId: contextId,
+          githubIntegration: repoIntegration,
+          repositories,
+        });
+      } catch (error) {
+        toast.error("Couldn't link repositories", {
+          description: error instanceof Error ? error.message : String(error),
+        });
+      }
     }
 
     if (withContextMd && trimmedDescription) {
@@ -350,6 +376,18 @@ export function CreateChannelModal({
 
             <DialogBody viewportClassName="flex flex-col gap-4">
               {descriptionField}
+              <Field>
+                <FieldLabel>Repositories (optional)</FieldLabel>
+                <RepositoriesField
+                  selected={repositories}
+                  integrationId={repoIntegration}
+                  disabled={busy}
+                  onChange={(nextRepositories, nextIntegration) => {
+                    setRepositories(nextRepositories);
+                    setRepoIntegration(nextIntegration);
+                  }}
+                />
+              </Field>
             </DialogBody>
 
             <DialogFooter>
